FARQXER-314 Backend con base de datos completo

+++ src/main/java/com/xeridia/xercoffeeshop/repository/PedidoRepository.java
@@ -3,5 +3,5 @@ package com.xeridia.xercoffeeshop.repository;
 import com.xeridia.xercoffeeshop.Pedido;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PedidoRepository extends CrudRepository<Pedido, Integer> {
+public interface PedidoRepository extends CrudRepository<Pedido, Long> {
 }
